GO-4459: Refactor into auth, object, search, utils
